[MU4] Fix (MinGW/gcc) compiler warnings

+++ src/framework/mpe/tests/multinotearticulationstest.cpp
@@ -241,21 +241,21 @@ TEST_F(MultiNoteArticulationsTest, CrescendoPattern)
 
     ArticulationPattern crescendoScope;
 
-    for (int i = 0; i <= dynamicSegmentsCount; ++i) {
+    for (size_t i = 0; i <= dynamicSegmentsCount; ++i) {
         ArticulationPatternSegment crescendoPattern;
         crescendoPattern.arrangementPattern = createArrangementPattern(HUNDRED_PERCENTS /*duration_factor*/, 0 /*timestamp_offset*/);
         crescendoPattern.pitchPattern = createSimplePitchPattern(0 /*increment_pitch_diff*/);
         crescendoPattern.expressionPattern = createSimpleExpressionPattern(dynamicLevelFromType(
-                                                                               DynamicType::Natural) + i * DYNAMIC_LEVEL_STEP
+                                                                               DynamicType::Natural) + static_cast<int>(i) * DYNAMIC_LEVEL_STEP
                                                                            / static_cast<dynamic_level_t>(dynamicSegmentsCount));
 
-        crescendoScope.emplace(25 * ONE_PERCENT * i, std::move(crescendoPattern));
+        crescendoScope.emplace(25 * ONE_PERCENT * static_cast<int>(i), std::move(crescendoPattern));
     }
 
-    for (int i = 0; i < noteCount; ++i) {
+    for (size_t i = 0; i < noteCount; ++i) {
         // [GIVEN] Crescendo articulation applied on the note
-        ArticulationData crescendoApplied(ArticulationType::Crescendo, crescendoScope, 25 * ONE_PERCENT * i,
-                                          25 * ONE_PERCENT * (i + 1), 0, dynamicLevelDiff);
+        ArticulationData crescendoApplied(ArticulationType::Crescendo, crescendoScope, 25 * ONE_PERCENT * static_cast<int>(i),
+                                          25 * ONE_PERCENT * (static_cast<int>(i) + 1), 0, dynamicLevelDiff);
         appliedArticulations[i].emplace(ArticulationType::Crescendo, crescendoApplied);
     }
